Restructure code for modularity
This is an initial commit to allow for more modularity in the code.  Next commit up will add another abstraction layer for storage into sandcastle.

This commit also removes all of the event-stream stuff and replaces it with just plain javascript.  I find it easier to read and understand what is going on.