[BACK-1758] Move data uploads/datasets into their own collection. (#186)
* Do all modifications of dataSets (datum.type == "upload") across old deviceData
and new deviceDataSets collection in a transaction. Unfortunately the callback
API is far too messy for this so some refactoring to use async/await was added.

updateDatumInternal was also refactored to be asynchronous and to get rid of
the recursive-like setTimeOut that is no longer needed as the transactions API
will already prevent concurrent updates and automatically retry for us.